pyproject.toml
[project]
name = "automaton-auditor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "astroid>=4.1.1",
    "docling>=2.75.0",
    "fpdf>=1.7.2",
    "gitingest>=0.3.1",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.9",
    "pdf2image>=1.17.0",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "pypdf>=6.7.3",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "reportlab>=4.4.10",
]
